Post Access Control
Posts cannot be edited or deleted by anyone other than the author, and the author must be logged in to do so.

+++ laravel/app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use DB;
 
 class PostsController extends Controller
 {
+	/**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+	
     /**
      * Display a listing of the resource.
      *
@@ -96,7 +106,17 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+		
+		// get id
         $post = Post::find($id);
+		
+		// check if correct id 
+		if(auth()->user()->id !== $post->user_id){
+			return redirect('posts.posts')->with('error', 'Unauthorised Page: Access Denied');
+		}
+		
+		
+		
 		return view('posts.edit')->with('post', $post);
     }
 
@@ -135,6 +155,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+		
+		// check if correct id 
+		if(auth()->user()->id !== $post->user_id){
+			return redirect('posts.posts')->with('error', 'Unauthorised Page: Access Denied');
+		}
+		
+		
 		$post->delete();
 		return redirect('/posts')->with('success', 'Post Removed');
     }
